Use catalina.base instead of catalina.home for log files to support multi-instance tomcat.
Tomcat allows multiple instances to be installed using the same binary. In debian, this is performed by the tomcat7-user package. In centos, this can be achieved by customizing the init script and copying a skeleton tomcat directory structure. In both cases, tomcat sets the CATALINA_HOME environment variable (catalina.home for java) to the location of the tomcat binary, and the CATALINA_BASE environment variable (catalina.base for java) to the location of the tomcat installation. catalina.sh describes that CATALINA_BASE will be set to CATALINA_HOME if not present.

    #   CATALINA_HOME   May point at your Catalina "build" directory.
    #
    #   CATALINA_BASE   (Optional) Base directory for resolving dynamic portions
    #                   of a Catalina installation.  If not present, resolves to
    #                   the same directory that CATALINA_HOME points to.

Since CATALINA_BASE contains the "dynamic portions" of the Catalina installation, this is the appropriate place to look for log files.

In the default oxauth installation, however, log4j is configured to create / look for the log files under CATALINA_HOME, which results in permissions / file not found errors if a tomcat instance is run.

This patch sets logfiles to be created / found under catalina.base rather than catalina.home. If CATALINA_BASE is not set, catalina will be setting it to CATALINA_HOME anyways, so there shouldn't be any damage done.