Update SamlPassportAuthenticator.py
The problem here is that instruction like `for provider in config:` where "config" is a dict will iterate through keys of it, not through values - and the subsequent instructions in the loop assume "provider" is a dict itself, while it's a string (because it's a key, like "idp1"), in fact: `equalsIgnoreCase(provider["enable"], "true")` The script fails to populate "registeredProviders" dict with values, as a result and fails.

Appropriate way to do that would be something like `equalsIgnoreCase(config[provider]["enable"], "true")`. I've updated that part of the script accordingly.