Docs: updated navigation items, removed duplicate files (#2737)
* docs: updated some navigation information

Commented out future files, renamed Docker Quick Start, added `setup.md`

* docs: removed Integration Guide from navigation

The integration guide is duplicated in the Recipes.

* docs: removed duplicate items from Navigation

* docs: removed duplicate files

The integration folder was duplicated in the Janssen Recipes folder but never deleted. The Casa doc is the only one with content, and this commit moves that to the correct location.

* docs: move setup file to Installation Guide

* docs: removed incorrectly copied content

When creating a new file, I grabbed content from a different file that shouldn't have been included.

Former-commit-id: 529c0eb0b195127b34b5353b6bd38ca8f7e380d6