fix: openapi spec version element (#6780)
* docs: add `version` tag to auth-server spec

Signed-off-by: ossdhaval <343411+ossdhaval@users.noreply.github.com>

* docs: add `version` tag to fido2 spec

We are adding blank versions as version info is captured in links shown on api SPEC when rendered on Swagger-UI

Signed-off-by: ossdhaval <343411+ossdhaval@users.noreply.github.com>

* docs: add `version` tag to SCIM spec

We are adding blank versions as version info is captured in links shown on api SPEC when rendered on Swagger-UI

Signed-off-by: ossdhaval <343411+ossdhaval@users.noreply.github.com>

* docs: add `version` tag API SPECs

Adding `OAS Version` as version since `version` element is mandatory but we don't want to maintain version info in this tag. Version info is already available in URL path.

Signed-off-by: ossdhaval <343411+ossdhaval@users.noreply.github.com>

---------

Signed-off-by: ossdhaval <343411+ossdhaval@users.noreply.github.com>
Signed-off-by: Mustafa Baser <mbaser@mail.com>