Update link to Authorization Challenge Custom Script Page
The Authorization Challenge Endpoint is a specialized endpoint in the Jans Authorization Server designed to support first-party native applications that need to obtain authorization codes without invoking a web browser. It enables a browserless OAuth 2.0 flow, which is particularly useful for mobile or desktop applications that cannot reliably open a browser for user interaction or rely on embedded authentication experiences.

Signed-off-by: ANEM BHARGAV <anembhargav02@gmail.com>